requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.4
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.50.0
chromadb==0.4.22
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
pymupdf==1.23.8
python-dotenv==1.0.0
tiktoken==0.7.0